requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
easyocr==1.7.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
idna==3.6
imageio==2.33.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python==4.8.1.78
orc==0.1.0
packaging==23.2
Pillow==10.1.0
pyclipper==1.3.0.post5
pyparsing==3.1.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
urllib3==2.1.0
